agent-driven: Use 32-bit arithmetic for days calculation

+++ components/calendar/src/tests/extrema.rs
@@ -68,7 +68,7 @@ fn nudge_duration_by_unit(
         DateDurationUnit::Months => duration.months = duration.months.saturating_add_signed(value),
         DateDurationUnit::Weeks => duration.weeks = duration.weeks.saturating_add_signed(value),
         DateDurationUnit::Days => {
-            duration.days = duration.days.saturating_add_signed(i64::from(value))
+            duration.days = duration.days.saturating_add_signed(value)
         }
     }
     duration
